pyproject.toml
[project]
name = "hackaton-2025"
version = "0.1.0"
description = "API para sistema de asistencia biométrica offline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.14.2",
]
